added animation after shuffeling
added the animations to shuffeling. will add animation on shuffleing on its own animation soon. it gives a bug when there is no if statment to check if shuffeling is happening. because it doesnt wait for the animation it does the animation and executes the next command while its still animating