Comments, renamed closeMe, fixed move counter
Improved the JSDoc comments to explain complicated code more.
renamed closeMe to closeWin to make the name of the function clearer.
I forgot to make the moves counter increment in the new move function so I fixed that too.
Also he mentioned using the default cursive so I included that in the styling. We can change it if yall want.